azs - major improvement in tmvn_MLE fitting
Used to be a gradient-descent based algorithm; now it starts with global optimization using differential evolution and only relies on gradient descent for the final polishing.
Also added a pre and post-processing step that standardizes the samples (and detection and truncation limits).
These two improvements lead to much better estimates at the cost of longer calculation time. Fitting shall not take more than a few minutes.