azs - DLC - allow the use of automatic P58/Hazus loss mapping with custom DBs
This required identifying the mapping method based on the assigned damage process and not the assigned DB. Consequently, automatic mapping will no longer work with a custom damage process.
We could introduce auto P58 and auto Hazus instead of just auto mapping to resolve this if it becomes an issue.