Remove `return` calls in `_perform_dmg_task`
`qnt_sample` is passed as an argument to `_perform_dmg_task` and in
the body of the function we perform operations that modify the
dataframe in palce, therefore we affect the object that was passed in
a "by reference" manner. Returning `qnt_sample` would imply that the
originally passed `qnt_sample` remains unchanged, so it's better to
follow the more explicit approach of not returning the object.