Modify unit conversion implementation
The previous code relied on `.loc` to determine the columns to scale.
This produced a pandas warning when the MultiIndex contained mixed
types.
This commit circumvents the warning by replacing the use of `.loc`
with numpy indexing. It requires keeping `nan` entries in the units
and removing them at a later stage, before they are returned, in order
to make dimensions match for the indexing operation.