azs - improve handling warning messages
Add a new option to Pelicun to separate controling warnings from providing verbose output. By default, all deprecation, future, and performance warnings are suppressed for the sake of clarity in output. A new attribute is added to the Logger class to handle warnings and the default setting is added to the default_settings.json.
This update allows us to remove the global warning suppression-related scripts from DLC and base.py and handle this behavior better through the options passed to Pelicun.