Implement performance group-specific capacity scaling
Adds the ability to scale the median capacity of specified performance
groups. Permitted operations include addition, subtraction,
multiplication, and division. Use of `eval` was avoided due to
security concerns. The median is directly modified at the stage of the
capacity RV definition, before sampling, to minimize codebase
expansion. Capacity scaling can only be applied to the
median (`theta_0`) of `normal` and `lognormal` RVs.  Includes a test
which may need to be expanded to maximize coverage, particularly for
possible exceptions.