Initial steps to improving `load_data`
This commit includes changes made in an attempt to simplify the code
in `load_data`. Currently the tests pass but some example use cases
fail. Also this implementation appears to be slower than the previous
code in certain cases. I am investigating the issues and will resolve
them in subsequent commits.