Improved `load_data` method
This commit includes a reqritten version of `load_data`.  Similar to
the prior version, it avoids transposing the dataframe, as this can be
a slow operation when changes are made to it.
The commit includes a new approach to inferring the data types of the
dataframe's contents. When `Units` are present, this typically forces
columns to have an `object` type. When removed, the data types of the
columns need to be reassigned. This is done with a new function in
`base`, `convert_dtypes`, and unit tests are added.
Unit conversions are handled by indiscriminately applying conversion
factors to all numeric data. Now it is handled in a vectorized
function, eliminating looping over units.
For cases where unit conversions are not required, the method
separates the unit row/column but does not apply conversion
factors. Instead of an empty list, now no conversion is specified by
`unit_conversion_factors` being None, which is more intuitive, and is
reflected in the updated docstring.