Default `truncation_limits` for all RV objects
This commit adds a default `truncation_limits` argument to all RV
objects, even for those that don't support it yet. The default is
`nan` in both directions, and for the RVs that don't support
truncation, a `NotImplementedError` is raised if the attribute is set
to a different value while instantiating the object.  This helps
eliminate the if/else logic in the `model` module code, where we had
to check if the distribution family is one for which we support
truncation and not add that argument if it doesn't. Now we rely on the
users to avoid specifying truncation limits for families that don't
support it, or they will be getting the `NotImplementedError` at
runtime.