azs - bugfix - check for valid dv_types and components when preparing sample for saving
When losses are calculated based on a damage table with DS0 in it, we
can have components that show up in the damage table, but end up being
excluded from the loss table since we do not save components with zero
losses - for now. This caused a problem when parsing units for DVs.