Preliminary refactoring - sample_size & changes in DL_calculation.py
In an effort to modularize the loss model, sample_size is turned into
an argument that has to be passed in `calculate` instead of having to
infer it from other assessment models. This small change evolved into
a larger refactoring effort to make sampling_size more explicit across
methods.

- The config schema was modified to include sampling information as
part of "Options".
- `sampling_size` is included in all `calcualte` methods as an
explicit argument and no longer has to be infered with `self._samnt.`
- Test suite updated

DL_calculation is modified to assign sample_size when calling
`.calculate` methods. Also:
- The `config` dictionary is used explictly instead of assigning
sub-dictionaries into variables. This helps understand the full path
to the retrieved item at any point in the code, instead of having to
remember what key corresponds to what variable.
- Improved some parts where there was strange syntax like
  `if out_config.get('Demand', None) is not None` ->
  `if 'Demand' is in out_config`