Remove inconsistent `return`
If there is no sample `aggregate_losses` returns None.
Aggregating losses without a repair sample is unusual and we need to
rethink what should be done in that case instad of returning None,
which is a bit unintuitive and turns the output type of the method to
optional, making things more complicated.