Introduce DamageModel_DR
When losses are driven by a loss function we use an EDP as input which
gets mapped to a loss value. Wanting to also assign a notion of damage
for such components, we utilize a "damage ratio", which ranges from 0
to 1, and is used as an itermediate measure to get the losses. Losses
are then obtained using this damage ratio as input.

EDP -> DR -> Loss

Since handling the damage assignment of such components is
fundamentally different, we split the existing DamageModel to
DamageModel_DS, dealing with the already supported damage state-driven
components, and DamageModel_DR, which will handle damage ratio-driven
components.