WIP - Expand DamageModel
This is a huge commit [^1].

With this commit `DamageModel.calculate` is back online and works for
DamageModel_DS (the new name of the already existing DamageModel).
Implementing the corresponding part for DamageModel_DR is currently in
progress.

[^1] Due to a bad habit of wanting all commits to be "working code."
This results in commits making massive changes that are impossible to
interpret. Going forward I will be committing more often with a Work
In Progress (WIP) tag, meaning that the code doesn't work/isn't
linted/isn't unit-tested, but is a partial edit in the process of
implementing a feature that requires making extensive changes.