WIP - Update loss model
In this commit everything works up to `calculate` (currently only for
the DS model).

The LossModel, which is contained in Assessment objects now uses the
loss map defined in its two loss models, dr_model and lf_model.

The decision variables that are considered for the loss calculations
are defined as an attribute before the calculation begins.

LossModel.calculate runs the calculate method of the two models.

For the Loss Function (LF) model, the part that loads the parameters
works. `_convert_marginal_params` had to be changed to do the unit
conversions properly, since in the case of loss functions "arg_units"
don't represent component quantities.