WIP - More modular code, sample size
Currently in the process of implementing the `calculate` method for
the loss function model. Since it will need to determine the demand
that drives the losses of each component, we need to make that part of
pelicun a bit more modular. So, the relevant code is being pulled off
of damage model and shoved into the damage model. This introduces the
opportunity to further de-tangle and modularize the code (relying less
on retrieving attributes from other models via `_asmnt`), which will
make testing easier.

Sample size:
Loss models can't have an independent sample size, since they need the
damage and demand model realizations. In general, the only way to run
an assessment, even when fewer models are used (e.g. no damage, or no
losses), the sample size has to be the same for all models involved.