Add warnings in a separate file too.
This commit configures the logger so that when a log file is
requested, another file with the suffix `_warnings` is added,
containing only the warnings. The warnings are still being placed in
the log file as well. The file name for the warnings file takes care
of any file extensions, even when there are periods in the file name
itself before the file extension. Extra spaces and color formatting is
removed from the warnings so that the log file doesn't contain escape
sequences used for coloring the terminal output.