Implement loss combination and validation logic for aggregate_losses.
- Enhanced `aggregate_losses` method to support non-additive loss
combinations.
- Implemented methods `_validate_input_replacement_thresholds` and
`_validate_input_loss_combination` for validating input dictionaries.
- Modified loss combination data files and added new test data files.
- Updated `test_loss_model.py` to include a new test for
`aggregate_losses` with loss combinations.
- Added validation scripts for 3D interpolation and pandas conversion
speed tests in `tests/validation/inactive`.
- Minor CSV format adjustments in `Wind_Flood_Hazus_HU_bldg.csv`.

Notes
-----
The loss combination parameter specification differs from what
discussed in the development meeting. It does not involve a pandas
DataFrame, but rather a numpy array directly. The domains are inferred
by the array itself using the first column and first row (and so on,
for higher-dimensional arrays), where the loss values of all but one
components need to be zero, so there is no need to attach an index to
each dimension of the array.

The implementation assumes combinations of M components using
M-dimensional combination arrays rather than being restricted to 2D or
3D arrays as separate cases. It makes use of sets, which are native
python objects, and set operations, which are very efficient. It also
utilizes caching to avoid repeated operations.

A great way to understand the logic is to follow the execution steps
of `test_aggregate_losses_combination`, ideally in debugging mode
which would allow variable inspection.