Remove unreachable code
We initialize `out_files` as [] and then check if the existing files
are in this empty list. They would never be members of the empty list,
so the deletion code is uncreachable.
This commit simply removes it. The expectation now is that existing
files would be overwritten if the output filename already exists for
some output, or left alone otherwise.