Extract config parsing method.
The idea behind this is to ensure the config is valid and contains
whatever it should (this is the goal with this method), but its
contents shouldn't be read into variables. Once the config is
validated/modified/extended, the rest of the code can use it directly
to get the needed information. Data retrieval from dictionaries is
really fast and there is practically no penalty in doing it multiple
times, so there is no need to extract data into variables.