Improve handling of `ds_id` in `map_ds`.
This commit addresses the last linter warning by bugbear, on not using
upper-scope variables as default arguments (since they are only
evaluated at the function definition).
Fortunately for us, this was the behavior we wanted, so the warning
didn't actually point to a bug. But to eliminate the warning
altogether and avoid this confusion, we can instead utilize `partial`
from the `functools` to create an on-the-spot partial evaluation of
`map_ds` with the current value of `ds_id`. At this point `map_ds`
could be moved outside of `assign_lsds` to reduce nesting, but it
also works as is.