Update the `sample` property of the `loss` model.
Update the `sample` property of the `loss` model to combine the
samples of the ds_model and lf_model. Backwards compatibility is still
maintained in the sense that legacy code will not result in a
lf_sample, and therefore the method will just return the ds_sample.
The non-existent `ds` MultiIndex level for the lf_model sample is
filled with 'N/A'. Currently this is not an issue because all
MultiIndex levels of the columns of the samples have `str` dtype.