Decision variable units.
This commit improves the way pelicun determines the right units to use
for each decision variable in an assessment, using a new loss model
attribute, `dv_units`.
If an existing sample is loaded, this commit forces the sample to
contain units.
When model parameters are loaded instead, we introduce checks to
ensure that only a single unit is used for all loss parameters
concerning a particular decision variable.
They are then inferred and stored in `dv_units` for use in loss aggregation.