Handle multiple logger case.
Improve implementation by ensuring an exception is logged by all
instantiated loggers.

This is achieved by creating a module-level LoggerRegistry class that
keeps track of all instantiated loggers and defines a `log_exception`
class method which is assigned to `sys.excepthook`. The class method
documents the exception to all registered logger objects in the form
of a warning.