Unreleased change log.
Update `Unreleased` change log.

This section will always be there to document the changes that are
worthy of mention in in the change log of the next release, so that we
don't have to painstakingly monitor the commit logs to write a change
log before making each release. When a release is made, the log will
be transferred from this page to the page associated with the new
version.