Update the built-in Hazus Hurricane assessment with flood models
This includes the following changes:

- Add support for multiple loss model files assigned to a single
built-in analysis type because Hazus Hurricane needs both wind and
flood loss data.

- Update loss map preparation to use the asset sample (instead of the
damage sample) so that we get all assigned components into the loss map.

- Leverage the new loss combination feature by making it available
through DL calculation and providing default combination matrix for
users of the Hazus Hurricane assessment. (The replacement threshold
configurations are also exposed, but not yet used in DL calculation)

- The Damage and Loss Model Library was also updated and several bugs
were fixed related to this feature. The corresponding changes are
described in the commits in that repo.

- Add the flood model file to the list of bundled resources