Replace auto-pop with pelicun config - Stage 1
This change is breaking compatibility with old R2D workflows. It
replaces the old auto-population routine in DLC with a pelicun config
that assumes all features are in the incoming AIM and only performs
mapping those to archetypes and preparing a configuration for the
Pelicun simulation.

The configuration is driven by a pelicun_config.py file that is expected
to be in the same folder as the DL models for the simulation.

For built-in models, we have shortcuts in the system where the user
specifies "HAZUS MH EQ IM" as the method and we automatically pull in
the path to the folder with the corresponding model and config files.

For user-defined models, we expect the user to provide a path to a
folder where the custom models are located. This is coming in as
custom_model_dir. When running from R2D, this dir is in the input_data
folder where the custom models are automatically copied before starting
the analysis.

After this update, there's no need for a separate auto-population script
input, since we expect the script to be in the custom dir and have a
standard name.

After this commit comes an updated link to the DLML where I've included
the new pelicun config and data schema files - see commits there for
details.