update dlml_resource handling
First, replace the file references (e.g., `damage_DB_FEMA_P58_2nd.csv`)
with separate references to a method (e.g., `FEMA P-58`) and the data
file within the method (e.g., `fragility.csv`). Only the method needs
to be looked up in the dlml_resource file and replaced with the path
that points to the corresponding folder within the Damage and Loss
Model Library. This provides more flexibility for users to utilize
various files from DLML and also yields a much leaner dlml_resource
file. When default databases are specified in the DL config, it is
sufficient to put the method name in, since the file type is trivial
from context (i.e., method for damage or for loss assessment)

Backwards compatibility is maintained. Folks who send in filenames
such as `damage_DB_FEMA_P58_2nd.csv` will receive a warning and we
convert those filenames to a method + filename combination using the
`legacy_names` dictionary in `file_io`.
I tried to move such compatibility checks and adjustments to the
`substitute_default_path` method in `file_io` for clarity and reduced
redundancy. In some cases, I had to keep a few in `assessment`.

The combination rules for Hazus Hurricane deserve a special mention:
these are loaded in the `calculate_loss` method in `assessment`. I
replaced the access to these with the new format that should work, but
it requires further testing.